Don’t run benchmark with mocha
