+++ package.json
@@ -7,7 +7,7 @@
     "test": "npm run standard && node_modules/.bin/mocha ./test/helpers/setup  test/marker-index.test.js",
     "tdd": "npm run standard && node_modules/.bin/electron-mocha --renderer --interactive --recursive ./test/helpers/setup  test/marker-index.test.js",
     "profile": "npm run standard && node_modules/.bin/electron-mocha --renderer --interactive --recursive ./test/helpers/setup  test/marker-index.profile.js",
-    "benchmark": "npm run standard && node_modules/.bin/mocha ./test/helpers/setup  test/marker-index.profile.js",
+    "benchmark": "npm run standard && node test/marker-index.benchmark.js",
     "prepublish": "npm run standard && npm run clean && npm run build",
     "standard": "node_modules/.bin/standard --recursive src test",
     "build": "node_modules/.bin/babel src --out-dir dist",
